# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "用户模块",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stores/getCode": {
            "post": {
                "summary": "获取验证码",
                "operationId": "App\\Http\\Controllers\\Store\\AuthController::getCode",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "store_mobile": {
                                        "description": "手机号",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{code:0,message:'ok'}"
                    }
                }
            }
        },
        "/api/stores/login": {
            "post": {
                "summary": "商家登录",
                "operationId": "App\\Http\\Controllers\\Store\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "shop_no": {
                                        "description": "店铺号（SP+手机号）",
                                        "type": "string"
                                    },
                                    "store_mobile": {
                                        "description": "手机号",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "登录密码",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{status:1（1.成功，0.失败）,msg:提示语}"
                    }
                }
            }
        },
        "/api/stores/register": {
            "put": {
                "summary": "商家注册",
                "operationId": "App\\Http\\Controllers\\Store\\AuthController::register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "store_mobile": {
                                        "description": "手机号",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "登录密码",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "确认密码",
                                        "type": "string"
                                    },
                                    "invite_code": {
                                        "description": "邀请码",
                                        "type": "string"
                                    },
                                    "sms_code": {
                                        "description": "验证码",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "{status:1（1.成功，0.失败）,msg:'提示语'}"
                    }
                }
            }
        },
...
    }
}